Introduce the UserAccess model
