+++ app/models/user.rb
@@ -11,11 +11,7 @@ class User < ApplicationRecord
     allowed: "allowed",
     denied: "denied"
   }, _prefix: true
-  enum access_level: {
-    viewer: "viewer",
-    manager: "manager",
-    power_user: "power_user"
-  }, _suffix: :access
+  enum access_level: UserAccess::LEVELS.map { |level, meta| [level, meta[:id].to_s] }.to_h, _suffix: :access
 
   belongs_to :organization, optional: true
   has_many :user_authentications
@@ -77,6 +73,13 @@ class User < ApplicationRecord
     :password,
     :authenticatable_salt,
     :invited_to_sign_up?, to: :email_authentication, allow_nil: true
+  delegate :accessible_organizations,
+    :accessible_facilities,
+    :accessible_facility_groups,
+    :can?,
+    :grant_access,
+    :access_tree,
+    :permitted_access_levels, to: :user_access, allow_nil: false
 
   after_destroy :destroy_email_authentications
 
@@ -86,6 +89,10 @@ class User < ApplicationRecord
 
   def email_authentication
     email_authentications.first
+  end
+
+  def user_access
+    UserAccess.new(self)
   end
 
   def registration_facility_id
@@ -183,31 +190,6 @@ class User < ApplicationRecord
   def resources
     user_permissions.map(&:resource)
   end
-
-  # #########################
-  # User Access (permissions)
-  #
-  def accessible_organizations(action)
-    return Organization.all if power_user?
-    accesses.organizations(action)
-  end
-
-  def accessible_facility_groups(action)
-    return FacilityGroup.all if power_user?
-    accesses.facility_groups(action)
-  end
-
-  def accessible_facilities(action)
-    return Facility.all if power_user?
-    accesses.facilities(action)
-  end
-
-  def can?(action, model, record = nil)
-    return true if power_user?
-    accesses.can?(action, model, record)
-  end
-  #
-  # #########################
 
   def destroy_email_authentications
     destroyable_email_auths = email_authentications.load
